requirement.txt
fastapi
uvicorn
pandas
pydantic
asyncio
httpx
python-dotenv
numpy
matplotlib
seaborn
ipykernel